# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIKARIR API",
        "description": "Dokumentasi API untuk aplikasi Sikarir — Sistem Informasi Karir & Pelatihan BLK",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/lowongan": {
            "get": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Ambil semua data daftar lowongan (paginate 15)",
                "operationId": "bfcbe9e605e81f4e86cd74a2cf06f56c",
                "responses": {
                    "200": {
                        "description": "Daftar lowongan berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Buat data lowongan baru",
                "operationId": "007bce9a1dbc0125441ab80d94fc29b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mitra_id",
                                    "judul"
                                ],
                                "properties": {
                                    "mitra_id": {
                                        "type": "integer"
                                    },
                                    "admin_id": {
                                        "type": "integer"
                                    },
                                    "judul": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "gaji": {
                                        "type": "number"
                                    },
                                    "lokasi": {
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lowongan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lowongan/{id}": {
            "get": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Ambil detail lowongan berdasarkan ID",
                "operationId": "a71b50c8f2519cdb72d8cbc180e116b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lowongan"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Update data lowongan",
                "operationId": "16799d15a95059b1858162bde736ba0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "gaji": {
                                        "type": "number"
                                    },
                                    "lokasi": {
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lowongan berhasil diupdate"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Hapus data lowongan",
                "operationId": "6c79bb0f331e5b1dc825e82e1bde295a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lowongan berhasil dihapus"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token yang diperoleh dari endpoint /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}